refactor(auth/signup): 회원가입 서비스 API 연결 및 타입 정리
- 이메일 인증 및 닉네임 체크 기능을 실제 API와 연결
- 불필요한 주석 제거 및 타입 정리

Related to #40
