+++ src/pages/(auth)/signup/services/index.ts
@@ -1,2 +1,1 @@
 export { signupService } from './signupService.ts';
-export type { SignupRequest, SignupPayload } from './signupService.ts';
