+++ src/pages/(auth)/signup/types/index.ts
@@ -1,2 +1,8 @@
 export * from './types';
 export type { SignupFormData } from '../schemas';
+export type {
+  SignupPayload,
+  EmailVerifySendPayload,
+  EmailVerifyConfirmPayload,
+  NicknameCheckPayload,
+} from './signup';
